+++ app/controllers/search_credentials_controller.rb
@@ -1,4 +1,6 @@
 class SearchCredentialsController < CredentialsController
+  before_action { head :not_found unless Rails.configuration.x.external_search_enabled }
+
   self.credential_class = SearchCredential
   self.validation_job = SearchCredentialValidationJob